added draw function
I wanted to be able to have continuous drawing instead of clicking each box div one at a time to change the colour. However, I wanted the user to be able to start and stop the continuous drawing. Therefore, I could not just use the mouseover event, I had to add a draw function to see whether the user had clicked on the grid.