Fixed grid size
Fixed the grid size so it is always 600px x 600px. I did this by removing the floor function on the length calculation, so no extra space is left behind in the grid. Also, set the box max width and height so the DOM doesn't change the length it is given.